refactor: rename provider to providers file, use common types for ProvidersMap
